Add files via upload
So, the first code was not working properly because I needed to find out how RFID cards actually work. What happens is that RFID cards are only meant as a sort of verification middle-man to then be able to access information from databases and files or tables. I was trying to shove everything into the RFID card or tag and that would not work because the capacity is so small. So what I then though of was making it so that the Admin in this case is the "database" since I don't have access to a database and it is out of the scope of this project. The admin can view the RFID's internal id, and create a file with the same id number as the tag/card. Then all of the information is stored into the text file and the necessary information is grabbed from the text file once the other occupations can verify themselves through a verification ID. Once they verify themselves, the card/tag can be scanned and the information they need to see is displayed. The program will only open the file whose name matches the RFID card's/tag's internal ID number. This is the simplest way for now on how to demonstrate how databases connect the information read from NFC/Credit Cards/etc. to Payment Service Providers and the such as everything is then taken care of through network communication.