Add files via upload
The WorkingGUI.py is the final file for the project. It takes the working code before GUI and uses Tkinter in order for there to be a GUI that displays the information for the Medic, Police Officer, and the Airport Security after scanning the RFID card/tag after their verification. The Admin also has some prompts that are done through the windows that Tkinter allows creation of but there are also parts of the Admin that will still be done through the terminal. I kept it like this due to the Admin really just being a human input database. However, most of the inputs for the Admin are through windows through Tkinter just because it's easier for me to input some arguments and decisions through the GUI than the terminal. The Admin is personal choice. However, the most important part for me was being able to see the correct information on a window whenever the occupations verified themselves and scanned the RFID. I tried several ways for there to be a prompt to scan the rfid card/tag after verification but the RFID functions took priority.