Add files via upload
Final Modification was for the save file text window to show that instead of the file being created again, it shows that the information was saved. I'm happy with the code :) I know some ideas might pop up in the future that I might want to add but I am really happy with how this project turned out and I hope it is enjoyed. I think the only thing that comes to mind that I would want to change is the fact that when it displays the text for the numeric id, I would have liked it to be able to be copied and pasted and for some reason it can't. For now I do not know if that is a limitation with Tkinter but it would make it easier for the Admin to put the file name correctly to just copy and paste. However, the code is fine as it would be a lot more efficient if done through a database and the numeric id would just be in it's own column in the database after being extracted so it is just a convenience I would have liked to have. Anyways, I am done for now. Enjoy!